some structral changes
