+++ sunpy/net/SPICE/Solo/solar_orbiter.py
@@ -3,9 +3,10 @@ import warnings
 import urllib.request
 
 from bs4 import BeautifulSoup
-
+from sunpy.net.base_client import BaseClient,QueryResponseTable
 from astropy.time import Time
-
+import sunpy.net._attrs as a
+from sunpy.net.SPICE import attrs as sa
 BASE_URL = "https://spiftp.esac.esa.int/data/SPICE/SOLAR-ORBITER/kernels/{}"
 
 KERNEL_TYPES = ["ck", "fk", "ik", "lsk", "pck", "sclk", "spk","mk"]
@@ -144,3 +145,80 @@ class SoloKernel:
             print("no match found for the search terms!")
 
         return filtered_kernel
+
+class SoloClient(BaseClient):
+
+    def __init__(self, mission,kernel_type):
+        self.mission = mission.upper()
+        self.kernel_type = kernel_type
+
+    def search(self,*query,**kwargs):
+        """
+        Search for SPICE kernels based on mission and other criteria.
+
+        Parameters:
+        - kernel: Type of kernel to search for (e.g., "ck", "ik", "sclk").
+        - instrument: Specific instrument for instrument kernels.
+        - version: Version of the kernel.
+        - start: Start date.
+        - end: End date.
+        - link: exact name of kernel
+
+        Returns:
+        - QueryResponseTable with matching kernels.
+        """
+        results = []
+
+        if self.mission == "SOLO":
+            solo_kernel = SoloKernel(self.kernel_type)
+            if isinstance(query,a.Time):
+                pass
+            if isinstance(query,sa.sensor):
+                pass
+            if isinstance(query,sa.Instrument):
+                pass
+            if isinstance(query,sa.link):
+                pass
+            filtered_kernels = solo_kernel.filter_kernels(**kwargs)
+
+            for index, link in filtered_kernels.items():
+                results.append({
+                    'Mission': self.mission,
+                    'Kernel': self.kernel_type,
+                    'Link': link,
+                    'Index': index
+                })
+
+        return QueryResponseTable(results)
+
+    def fetch(self, query_results, path=None, **kwargs):
+        """
+        Fetch the selected kernels.
+
+        Parameters:
+        - query_results: Results returned from the search method.
+        - path: Destination path for downloaded files.
+
+        Returns:
+        - List of paths to downloaded files.
+        """
+        if path is None:
+            path = './{file}'
+
+        downloaded_files = []
+
+        for result in query_results:
+            kernel_type = result['Kernel']
+
+            solo_kernel = SoloKernel(kernel_type)
+            index = result['Index']
+            downloaded_files.append(solo_kernel.download_by_index(index, destination=path))
+
+        return downloaded_files
+
+    @staticmethod
+    def _can_handle_query(*query):
+        """
+        Check if this client can handle the given query.
+        """
+        return True
